Move backup related logic into utils folder
